bug fix - starting/optimal/max and buffer_level changed from int to int64_t
buffer_level in VP8_COMP and starting_buffer_level, optimal_buffer_level
and maximum_buffer_size in VP8_CONFIG changed from int to int64_t
to avoid potential crash issues for larger target bit rates.

+++ vp8/common/onyx.h
@@ -145,9 +145,9 @@ extern "C"
         int over_shoot_pct;
 
         // buffering parameters
-        int starting_buffer_level;  // in seconds
-        int optimal_buffer_level;
-        int maximum_buffer_size;
+        int64_t starting_buffer_level;  // in seconds
+        int64_t optimal_buffer_level;
+        int64_t maximum_buffer_size;
 
         // controlling quality
         int fixed_q;
